requirements.txt
absl-py==2.0.0
addict==2.4.0
appdirs==1.4.4
basicsr==1.4.2
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.1.1
cupy-cuda12x==12.3.0
cycler==0.12.1
docker-pycreds==0.4.0
einops==0.7.0
fastrlock==0.8.2
filelock==3.13.1
fonttools==4.46.0
fsspec==2023.12.1
future==0.18.3
gitdb==4.0.11
GitPython==3.1.40
google-auth==2.25.1
google-auth-oauthlib==1.0.0
grpcio==1.59.3
huggingface-hub==0.19.4
idna==3.6
imageio==2.33.0
imgaug==0.4.0
importlib-metadata==7.0.0
importlib-resources==6.1.1
joblib==1.3.2
kiwisolver==1.4.5
lazy_loader==0.3
lmdb==1.4.1
lpips==0.1.4
Markdown==3.5.1
MarkupSafe==2.1.3
matplotlib==3.7.4
networkx==3.1
numpy==1.23.2
oauthlib==3.2.2
opencv-python==4.5.4.60
packaging==23.2
Pillow==10.1.0
pip==23.3.1
platformdirs==4.1.0
protobuf==4.25.1
psutil==5.9.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyparsing==3.1.1
pyrtools==1.0.2
python-dateutil==2.8.2
PyWavelets==1.4.1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
safetensors==0.4.1
scikit-image==0.17.2
scikit-learn==1.3.2
scipy==1.9.1
sentry-sdk==1.38.0
setproctitle==1.3.3
setuptools==68.0.0
shapely==2.0.2
shortuuid==1.0.9
six==1.16.0
smmap==5.0.1
tb-nightly==2.14.0a20230808
tensorboard-data-server==0.7.2
tensorboard-plugin-wit==1.8.1
thop==0.1.1.post2209072238
threadpoolctl==3.2.0
tifffile==2023.7.10
timm==0.9.12
tomli==2.0.1
torch==1.8.1+cu111
torchaudio==0.8.1
torchvision==0.9.1+cu111
tqdm==4.66.1
typing_extensions==4.8.0
urllib3==2.1.0
wandb==0.16.1
wcwidth==0.2.5
Werkzeug==3.0.1
wheel==0.41.2
yapf==0.40.2
zipp==3.17.0
